Update repository structure
